added process scheduler

+++ src/linked_list.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <iostream>
 
 template <class T>
 class Node
@@ -14,21 +15,21 @@ class LinkedList
 {
   Node<T> *firstItem;
 
-  Node<T> *getNodeAtPosition(int position);
+  Node<T> *getNodeAtPosition(ulong position);
 
 public:
   LinkedList() : firstItem(nullptr) {}
 
   void unshift(T *value);
-  void insertAtPosition(int position, T *value);
+  void insertAtPosition(ulong position, T *value);
   void push(T *value);
 
   T *shift();
-  T *removeAt(int position);
+  T *removeAt(ulong position);
   T *pop();
 
-  T *get(int position);
-  int size();
+  T *get(ulong position);
+  size_t size();
   T *last();
 
   ~LinkedList();
